+++ src/deadline/blender_submitter/addons/deadline_cloud_blender_submitter/open_deadline_cloud_dialog.py
@@ -16,6 +16,7 @@ from . import blender_utils as bu
 from . import sanity_checks as sc
 from . import scene_settings_widget as ssw
 from . import template_filling as tf
+from . import ocio_utils as ocio
 from ._version import version
 from ._version import version_tuple as adaptor_version_tuple
 
@@ -60,6 +61,17 @@ def create_deadline_dialog(parent=None) -> SubmitJobToDeadlineDialog:
     else:
         settings.output_path = os.path.dirname(bpy.context.blend_data.filepath)
 
+    if bpy.path.basename(bpy.context.scene.render.filepath):
+        settings.output_file_prefix = bpy.path.basename(bpy.context.scene.render.filepath)
+
+    # Read the user's preferences to set GPU settings.
+    settings.enable_gpu = bpy.context.scene.cycles.device == "GPU"
+
+    if bpy.context.preferences.addons["cycles"].preferences.compute_device_type != "NONE":
+        settings.gpu_device = bpy.context.preferences.addons[
+            "cycles"
+        ].preferences.compute_device_type
+
     # Load and set sticky settings, if any.
     settings.load_sticky_settings(settings.project_path)
 
@@ -68,6 +80,12 @@ def create_deadline_dialog(parent=None) -> SubmitJobToDeadlineDialog:
 
     # Set auto-detected attachments.
     auto_detected_attachments = _get_auto_detected_assets(settings.project_path)
+
+    # If the user has an OCIO config file set, we will have to upload any directories referenced by that file as job attachments.
+    if ocio.get_ocio_path():
+        ocio_config = ocio.get_ocio_config(ocio.get_ocio_path())
+        for path in ocio.get_ocio_referenced_dirs(ocio_config):
+            auto_detected_attachments.input_directories.add(str(path))
 
     # Set regular attachments.
     attachments = AssetReferences(
@@ -127,11 +145,38 @@ def _create_bundle(
         requirements: The host requirements. There are only passed if the user has specified custom host requirements in the UI. If "all available worker hosts" are selected, this is None.
         purpose: The purpose of the job bundle.
     """
+    _create_bundle_internal(
+        widget, job_bundle_dir, settings, queue_parameters, asset_references, requirements, purpose
+    )
+
+
+def _create_bundle_internal(
+    widget: SubmitJobToDeadlineDialog,
+    job_bundle_dir: str,
+    settings: tf.BlenderSubmitterUISettings,
+    queue_parameters: list[dict[str, Any]],
+    asset_references: AssetReferences,
+    requirements: Optional[dict[str, Any]] = None,
+    purpose: Optional[str] = None,
+    prompt_for_saving=True,
+) -> None:
+    """Create and write Deadline job bundle files to the given directory. Also save sticky settings.
+    The internal variation exists so that the functionality can be accessed without a UI.
+
+    Args:
+        widget: The submitter dialog widget.
+        job_bundle_dir: The directory to write the job bundle files to.
+        settings: The job settings.
+        queue_parameters: The queue parameters.
+        asset_references: The asset references.
+        requirements: The host requirements. There are only passed if the user has specified custom host requirements in the UI. If "all available worker hosts" are selected, this is None.
+        purpose: The purpose of the job bundle.
+    """
     # Make sure the output path used for submission is absolute
     settings.output_path = bpy.path.abspath(settings.output_path)
 
     # Run sanity checks on submission
-    sc.run_sanity_checks(settings)
+    sc.run_sanity_checks(settings, prompt_for_saving)
 
     renderable_cameras = bu.get_renderable_cameras(settings.scene_name)
 
@@ -140,7 +185,7 @@ def _create_bundle(
         frame_range=bu.get_frames(),
         renderable_camera_names=renderable_cameras,
         output_directories=settings.output_path,
-        output_file_prefix="image_###.png",
+        output_file_prefix=settings.output_file_prefix,
         image_resolution=(
             bpy.context.scene.render.resolution_x,
             bpy.context.scene.render.resolution_y,