+++ src/deadline/blender_submitter/addons/deadline_cloud_blender_submitter/template_filling.py
@@ -1,11 +1,6 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 
-"""Functions to fill Blender job templates.
-
-This file should not have a dependency on `bpy`, so it can be unit-tested.
-This implies that it should not import modules under `deadline_cloud_blender_submitter`,
-since that module imports `bpy` in `__init__.py`.
-"""
+"""Functions to fill Blender job templates."""
 
 from __future__ import annotations
 
@@ -38,6 +33,8 @@ class BlenderSubmitterUISettings:
     submitter_name: str = field(default="Blender")
     renderer_name: str = field(default="cycles", metadata={"sticky": True})
     scene_name: str = field(default="Scene", metadata={"sticky": True})
+    enable_gpu: bool = field(default=False, metadata={"sticky": True})
+    gpu_device: str = field(default="None", metadata={"sticky": True})
 
     ui_group_label: str = field(default="Blender Settings")
 
@@ -50,9 +47,13 @@ class BlenderSubmitterUISettings:
     # Paths and files.
     project_path: str = field(default="")
     output_path: str = field(default="")
+    output_file_prefix: str = field(default="output_####")
     input_filenames: list[str] = field(default_factory=list, metadata={"sticky": True})
     input_directories: list[str] = field(default_factory=list, metadata={"sticky": True})
     output_directories: list[str] = field(default_factory=list, metadata={"sticky": True})
+
+    # OCIO config file is specified here.
+    ocio_config_path: str = field(default="")
 
     # Layers settings.
     view_layer_selection: str = field(default="ViewLayer", metadata={"sticky": True})
@@ -206,6 +207,10 @@ def fill_job_template(
         }
     )
 
+    # If an OCIO config file is set, we will have to add a job parameter and environment for it.
+    if settings.ocio_config_path:
+        _add_ocio_template_data(job_template)
+
     # For each layer, create a step based on the default step template.
     steps = []
     default_step_template = job_template["steps"][0]
@@ -258,6 +263,26 @@ def fill_job_template(
         job_template["jobEnvironments"].append(override_environment["environment"])
 
     return job_template
+
+
+def _add_ocio_template_data(job_template: dict):
+    job_template["parameterDefinitions"].append(
+        {
+            "name": "OCIOConfigPath",
+            "type": "PATH",
+            "objectType": "FILE",
+            "dataFlow": "IN",
+            "description": "The colour management info to render with.",
+        }
+    )
+
+    if "jobEnvironments" not in job_template:
+        job_template["jobEnvironments"] = []
+
+    # Insert at index 0 so that OCIO is set before Blender runs
+    job_template["jobEnvironments"].insert(
+        0, {"name": "Set OCIO Path", "variables": {"OCIO": "{{Param.OCIOConfigPath}}"}}
+    )
 
 
 def _fill_step_template(
@@ -361,6 +386,7 @@ def get_parameter_values(
         "OutputDir": layer_settings.output_directories,
         "RenderScene": layer_settings.scene_name,
         "RenderEngine": layer_settings.renderer_name,
+        "GPUDevice": settings.gpu_device,
     }
 
     if layer_settings.frames_parameter_name:
@@ -380,6 +406,10 @@ def get_parameter_values(
             "The following queue parameters conflict with the Blender job parameters:\n"
             + f"{', '.join(overlap)}"
         )
+
+    # If specified, add the parameter value for OCIO here
+    if settings.ocio_config_path:
+        params.append({"name": "OCIOConfigPath", "value": settings.ocio_config_path})
 
     # If we're overriding the adaptor with wheels, remove the adaptor from the Packages
     if settings.include_adaptor_wheels: