+++ test/integ/helpers/test_runners.py
@@ -1,0 +1,69 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+
+import subprocess
+import yaml
+import json
+
+from pathlib import Path
+from typing import Any
+
+
+def run_command(args: list[str]) -> subprocess.CompletedProcess[bytes]:
+    output = subprocess.run(args, capture_output=True)
+
+    print(f"Ran the following: {' '.join(output.args)}")
+    print(f"\nstdout:\n\n{output.stdout.decode('utf-8', errors='replace')}")
+    print(f"\nstderr:\n\n{output.stderr.decode('utf-8', errors='replace')}")
+
+    return output
+
+
+def run_blender_submitter_test(
+    blender_location: Path, test_script_location: Path, *additional_args
+) -> subprocess.CompletedProcess[bytes]:
+    args = [
+        str(blender_location),
+        "--background",
+        "--python",
+        str(test_script_location),
+        # Don't allow automatic running of scripts in the scene
+        # See https://docs.blender.org/manual/en/latest/advanced/scripting/security.html
+        "--disable-autoexec",
+        "--python-use-system-env",
+        "--python-exit-code",
+        "1",
+        "--addons",
+        "deadline_cloud_blender_submitter",
+    ]
+
+    if additional_args:
+        args.extend(["--", *additional_args])
+
+    return run_command(args)
+
+
+def is_valid_template(template_location: Path) -> bool:
+    output = run_command(["openjd", "check", str(template_location), "--output", "json"])
+
+    output_json = json.loads(output.stdout)
+
+    return output_json["status"] == "success"
+
+
+def run_adaptor_test(template_path: Path, job_params: dict[str, Any]) -> None:
+    with open(template_path) as f:
+        template = yaml.safe_load(f)
+
+    for step in template["steps"]:
+        output = run_command(
+            [
+                "openjd",
+                "run",
+                str(template_path),
+                "--step",
+                step["name"],
+                "--job-param",
+                json.dumps(job_params),
+            ]
+        )
+        assert output.returncode == 0