+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## 0.5.1 (2025-01-14)
+
+### Bug Fixes
+* Fixes published release artifacts
+
+
+## 0.5.0 (2025-01-13)
+
+
+### Features
+* Enable GPU rendering from submitter ([`7a4f8bc`](https://github.com/aws-deadline/deadline-cloud-for-blender/commit/7a4f8bc28f55059b77a6f6d31baa39766fe56c98))
+
+### Bug Fixes
+* Add OCIO configuration support in SMF ([`3da609b`](https://github.com/aws-deadline/deadline-cloud-for-blender/commit/3da609b425cb9eab827b87fc0ff5b1179d6e7854))
+
+## 0.4.5 (2024-12-04)
+
+
+### Features
+* **adaptor**: Add BLENDER_EXECUTABLE environment variable (#156) ([`234c086`](https://github.com/aws-deadline/deadline-cloud-for-blender/commit/234c086d3b2626902aba76268f15f3c5e906953d))
+
+### Bug Fixes
+* Apply output file prefixes to submitter GUI ([`876348a`](https://github.com/aws-deadline/deadline-cloud-for-blender/commit/876348ac5eb8aec11b784dfe83e7ca46570c4793))
+
 ## 0.4.4 (2024-11-21)